pyproject.toml
[project]
name = "credit-score-modelling"
version = "0.1.0"
description = "A Python package for Credit Scorecard generation using OptBinning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.61.0",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "joblib>=1.5.1",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.5",
    "mlflow>=3.2.0",
    "optbinning>=0.20.1",
    "ortools<9.12.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pyprojroot>=0.3.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "streamlit>=1.48.0",
    "uvicorn>=0.35.0",
]
